+++ packages/rust-bindings/src/aux_functions.rs
@@ -170,7 +170,8 @@ pub fn get_credential_deployment_details_aux(
     Ok(response.to_string())
 }
 
-pub fn deserialize_credential_deployment_details_aux(input: &str) -> Result<String> {
+/// Given the bytes of a credential deployment (/AccountCredentialMessage), deserialize it and return as json.
+pub fn deserialize_credential_deployment_aux(input: &str) -> Result<String> {
     let credential_message: AccountCredentialMessage<Bls12, ExampleCurve, AttributeKind> = match crypto_common::from_bytes(&mut hex::decode(input)?.as_slice()) {
         Ok(c) => c,
         Err(e) => bail!(format!("{}", e))